refactor(capi): isolate allocator
